requirements.txt
tenable
pandas
xlsxwriter
python-dotenv
GitPython
